refactor: 서버 전수조사에서 나온 정리 항목을 반영한다
- 나간 채팅방의 알림, 고정 토글이 성공하고 고정 개수까지 차지하던 것을 고친다
- 공통 오류 응답에 429, 503을 더하고 생성 문서와 타입을 다시 만든다
- 회원 목록과 랭킹의 같은 조립 코드를 요약 서비스로 올린다
- 피드 목록이 닉네임 때문에 돌던 요약 조회를 목록 쿼리에 합친다
- 닉네임 검색이 계산해 놓고 버리던 즐겨찾기 서브쿼리를 걷어낸다
- 양방향 차단 확인을 한 쿼리로 합치고 방향별 검증을 리포지토리 테스트로 옮긴다
- operationId 충돌 여섯 쌍에 이름을 붙여 문서의 _1 접미사를 없앤다
- 어드민이 도메인 리포지토리를 직접 쓰던 곳과 프로젝션 위치를 규칙에 맞춘다
- 죽은 label 네 개와 SuspensionDetail, 미사용 테스트 의존성 두 개를 지운다
- 나이 계산, 회원 존재 확인, 비밀번호 처리, 푸시 우선순위를 공용 헬퍼로 모은다
- 쓰지 않는 포인트 내역 인덱스를 내리고 사진 업로드 제약 이름을 규칙에 맞춘다
- 비동기 실행기 설정이 웹소켓 때문에 필수임을 테스트로 고정한다

+++ server/src/main/kotlin/com/blueoauld/server/domain/member/service/MemberListCursor.kt
@@ -12,7 +12,7 @@ object MemberListCursor {
         return (orderValue.toDoubleOrNull() ?: return null) to (memberId.toLongOrNull() ?: return null)
     }
 
-    fun encodeRanking(likeCount: Int, locatedAt: Long, memberId: Long) =
+    fun encodeRanking(likeCount: Long, locatedAt: Long, memberId: Long) =
         listOf(likeCount, locatedAt, memberId).joinToString(SEPARATOR)
 
     fun decodeRanking(cursor: String?): Triple<Long, Long, Long>? {
